+++ src/main/java/com/bizcub/bedrockHotbar/mixin/ContextualBarRendererMixin.java
@@ -9,9 +9,9 @@ import org.spongepowered.asm.mixin.injection.invoke.arg.Args;
 
 /*? >=1.21.6 {*///~ contextual_bar
 import com.mojang.blaze3d.platform.Window;
-import net.minecraft.client.gui.contextualbar.ContextualBarRenderer;
+import net.minecraft.client.gui.contextualbar.ContextualBar;
 
-@Mixin(ContextualBarRenderer.class) //~ !contextual_bar
+@Mixin(ContextualBar.class) //~ !contextual_bar
 public interface ContextualBarRendererMixin {
 
     @Redirect(method = "top", at = @At(value = "INVOKE", target = "Lcom/mojang/blaze3d/platform/Window;getGuiScaledHeight()I"))
@@ -27,7 +27,7 @@ public interface ContextualBarRendererMixin {
 }
 
 //?} <=1.21.5 {
-/*import net.minecraft.client.gui.Gui;
+/*import net.minecraft.client.gui.Hud;
 
-@Mixin(Gui.class)
+@Mixin(Hud.class)
 public interface ContextualBarRendererMixin {}*///?}