+++ src/main/java/com/bizcub/bedrockHotbar/platform/Fabric.java
@@ -11,7 +11,7 @@ public class Fabric implements ClientModInitializer {
 
     @Override
     public void onInitializeClient() {
-        if (Compat.isClothConfigLoaded()) Main.init();
+        Main.init();
     }
 
     public static class ModMenu implements ModMenuApi {