+++ CHANGELOG.md
@@ -1,1 +1,2 @@
-* 1.21.11 update
+* Now the actionbar is also offset along with other elements.
+* As much duplicate code as possible has been removed.